Move: Script and Interpreter into subfolder
