+++ lib/script/index.js
@@ -1,0 +1,3 @@
+module.exports = require('./script');
+
+module.exports.Interpreter = require('./interpreter');
