+++ lib/script/interpreter.js
@@ -3,14 +3,13 @@
 var _ = require('lodash');
 
 var Script = require('./script');
-var Opcode = require('./opcode');
-var BN = require('./crypto/bn');
-var Hash = require('./crypto/hash');
-var BufferReader = require('./encoding/bufferreader');
-var BufferWriter = require('./encoding/bufferwriter');
-var Signature = require('./crypto/signature');
-var PublicKey = require('./publickey');
-var Transaction = require('./transaction');
+var Opcode = require('../opcode');
+var BN = require('../crypto/bn');
+var Hash = require('../crypto/hash');
+var BufferReader = require('../encoding/bufferreader');
+var BufferWriter = require('../encoding/bufferwriter');
+var Signature = require('../crypto/signature');
+var PublicKey = require('../publickey');
 
 /**
  * Bitcoin transactions contain scripts. Each input has a script called the
@@ -1056,6 +1055,7 @@ ScriptInterpreter.prototype.step = function() {
  * Translated from bitcoind's VerifyScript
  */
 ScriptInterpreter.prototype.verify = function(scriptSig, scriptPubkey, tx, nin, flags) {
+  var Transaction = require('../transaction');
   if (_.isUndefined(tx)) {
     tx = new Transaction();
   }
